fix: px->rem, props to searchcard, useMemo at getCity
