+++ package.json
@@ -31,6 +31,7 @@
     "react-query": "^3.39.3",
     "react-router": "^6.20.0",
     "react-router-dom": "^6.19.0",
+    "rollup": "^4.6.0",
     "vite": "^5.0.0"
   },
   "devDependencies": {
