+++ src/components/search/TourItem.jsx
@@ -1,17 +1,21 @@
+import { useMemo } from 'react'
 import { Link } from 'react-router-dom'
 import { Box, Typography } from '@mui/material'
 
 function TourItem({ attraction }) {
-  const [city] = attraction?.address?.split(' ')
+  const city = useMemo(
+    () => (attraction?.address ? attraction.address.split(' ')[0] : ''),
+    [attraction],
+  )
 
   return (
     // 클릭시 여행지 디테일 페이지로 이동
     <Link>
       <Box className="h-29 group relative mb-6 flex overflow-hidden border border-solid border-gray-200 bg-white hover:border-blue-200 hover:shadow-lg hover:shadow-blue-100">
-        <Box className="h-[126px] w-[126px] overflow-hidden">
+        <Box className="h-[8rem] w-[8rem] overflow-hidden">
           <img
             src="https://p1.pxfuel.com/preview/778/373/101/nature-landscape-rocks-formation-cave-beauty.jpg"
-            className="h-[126px] w-[126px] object-cover duration-1000 group-hover:scale-125 "
+            className="h-[8rem] w-[8rem] object-cover duration-1000 group-hover:scale-125 "
             alt="Nature"
           />
         </Box>
@@ -45,7 +49,7 @@ function TourItem({ attraction }) {
             <Typography
               variant="body1"
               className="ml-2"
-              style={{ marginLeft: '8px', display: 'inline' }}
+              style={{ marginLeft: '0.5rem', display: 'inline' }}
             >
               {attraction.reviewCount}건의 리뷰
             </Typography>
